feat(web): derive stat formatting from content
