+++ packages/web/src/translation/effects/formatters/stat/index.ts
@@ -1,11 +1,38 @@
 import { STATS } from '@kingdom-builder/contents';
-import { increaseOrDecrease, signed } from '../../helpers';
+import { gainOrLose, increaseOrDecrease, signed } from '../../helpers';
 import { registerEffectFormatter } from '../../factory';
-import { applyStatAddFormatter } from './registry';
 
 registerEffectFormatter('stat', 'add', {
-  summarize: (eff, ctx) => applyStatAddFormatter(eff, ctx, 'summarize'),
-  describe: (eff, ctx) => applyStatAddFormatter(eff, ctx, 'describe'),
+  summarize: (eff) => {
+    const key = eff.params?.['key'] as string;
+    const stat = STATS[key as keyof typeof STATS];
+    const icon = stat?.icon || key;
+    const amount = Number(eff.params?.['amount']);
+    const format = stat?.addFormat;
+    const prefix = format?.prefix || '';
+    if (format?.percent) {
+      const pct = amount * 100;
+      return `${prefix}${icon}${signed(pct)}${pct}%`;
+    }
+    return `${prefix}${icon}${signed(amount)}${amount}`;
+  },
+  describe: (eff) => {
+    const key = eff.params?.['key'] as string;
+    const stat = STATS[key as keyof typeof STATS];
+    const label = stat?.label || key;
+    const icon = stat?.icon || '';
+    const amount = Number(eff.params?.['amount']);
+    const format = stat?.addFormat;
+    const prefix = format?.prefix || '';
+    if (format?.percent) {
+      const pct = Math.abs(amount * 100);
+      return `${increaseOrDecrease(amount)} ${icon}${label} by ${pct}%`;
+    }
+    if (prefix) {
+      return `${increaseOrDecrease(amount)} ${prefix}${icon} by ${Math.abs(amount)}`;
+    }
+    return `${gainOrLose(amount)} ${Math.abs(amount)} ${icon} ${label}`;
+  },
 });
 
 registerEffectFormatter('stat', 'add_pct', {
@@ -46,8 +73,3 @@ registerEffectFormatter('stat', 'add_pct', {
     return `${increaseOrDecrease(0)} ${icon}${label}`;
   },
 });
-
-import './default';
-import './maxPopulation';
-import './absorption';
-import './growth';
